requirements.txt
openpyxl>=3.1.2
python-dotenv>=1.0.1
typer>=0.12.3
rich>=13.7.1
openai>=1.40.0
pymupdf>=1.24.9
python-docx>=1.1.2
streamlit
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
